alfir_project/alfir_site/app/requirements.txt
Django==4.2.13
djangorestframework==3.15.1
gunicorn==19.7.1
psycopg2-binary==2.9.9